+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 #### Bugs
 * Fix #3014: Resync Future is canceled and resync executor is shutdown on informer stop
+* Fix #2978: Fix SharedInformer NPE on initial requests while syncing
 * Fix #2989: serialization will generate valid yaml when using subtypes
 * Fix #2991: reduced the level of ReflectWatcher event received log
 * Fix #2992: allowing Watch auto-reconnect for shared informers
@@ -15,6 +16,7 @@
 * Fix #3027: fix NPE when sorting events in KubernetesResourceUtil
 
 #### Improvements
+* Fix #2788: Support FIPS mode in kubernetes-client with BouncyCastleFipsProvider
 * Fix #2910: Move crd-generator tests from kubernetes-itests to kubernetes-tests
 * Fix #3005: Make it possible to select which CRD version is generated / improve output